moved polygon details
